Explicitly defined servlet load order in web.xml
It looks like the problem we had with tomcat was related to the load-on-startup parameter used in web.xml.Because it's up to the servlet container (Jetty or Tomcat) which order it loads servlets that have the same <load-on-startup/> value, we noticed the following variation on loading:

Jetty: 1. Monkey Server, 2. jersey-servlet
Tomcat: 1. jersey-servlet , 2. Monkey Server

So we explicitly set a larger number for jersey-servlet to ensure same load order in Tomcat as in Jetty.