Trim boolean configuration before parsing
Boolean.parseBoolean() is strict with respects to whitespace.  When following the install guide,
I ended up with whitespace in my configuration.  "True " was parsed as false.

Rather than fix the configuration instructions, it is much easier and more friendly to
simply trim the value before parsing it, i.e. be tolerant in what we accept.