Refactor reading of chaos monkey properties and defaults values
Refactoring out the reading of 'enabled', 'maxTerminationsPerDay', and 'probability' into a separate method. The method also handles reading the default values from the group level if one doesn't exist at the named group level. The reason for the refactor is to override the way defaults are read internally at netflix.