Set region when using SES.
We ran into the issue where sending mail kept failing, even after we enabled production access for SES and verified the sender email address. The issue turned out that by default SimianArmy was using us-east-1 for SES, but the region we were operating in (and had production access/verified emails) was us-west-2.

It might be more sensible to use a property for controlling what region the SES client uses, but this works for us.